servicehub-mcp/pyproject.toml
[project]
name = "servicehub"
version = "0.1.0"
description = "Automotive Service Hub Portal + MCP Server on Databricks Apps"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "fastmcp>=2.12.5",
    "uvicorn>=0.34.0",
    "databricks-sdk>=0.60.0",
    "psycopg[binary,pool]>=3.1.0",
    "pydantic>=2",
    "mcp[cli]>=1.14.0",
]

[project.scripts]
servicehub = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]
